Create flights scheduler.
